requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.2
python-multipart==0.0.6
structlog==23.2.0
